+++ src/workerd/api/gpu/gpu-shader-module.c++
@@ -3,6 +3,7 @@
 //     https://opensource.org/licenses/Apache-2.0
 
 #include "gpu-shader-module.h"
+#include <workerd/io/io-context.h>
 
 namespace workerd::api::gpu {
